egn_api/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
google-cloud-pubsub==0.45.0
